+++ server/tests/backends/pandas_executor/test_pandas_executor_steps.py
@@ -4,11 +4,11 @@ from io import StringIO
 import geopandas as gpd
 import pandas as pd
 import pytest
-from toucan_connectors.common import nosql_apply_parameters_to_query
 
 from tests.utils import assert_dataframes_equals, get_spec_from_json_fixture, retrieve_case
 from weaverbird.backends.pandas_executor import execute_pipeline
 from weaverbird.pipeline import PipelineWithVariables
+from weaverbird.utils.toucan_connectors import nosql_apply_parameters_to_query_with_errors
 
 test_cases = retrieve_case("pandas_executor", "pandas")
 
@@ -30,7 +30,9 @@ def test_pandas_execute_pipeline(case_id, case_spec_file_path, available_variabl
 
     steps = spec["step"]["pipeline"]
     steps.insert(0, {"name": "domain", "domain": "in"})
-    pipeline = PipelineWithVariables(steps=steps).render(available_variables, nosql_apply_parameters_to_query)
+    pipeline = PipelineWithVariables(steps=steps).render(
+        available_variables, nosql_apply_parameters_to_query_with_errors
+    )
 
     domains = {"in": df_in, **dfs_in_others}
     result = execute_pipeline(pipeline, domain_retriever=lambda x: domains[x])[0]