+++ server/tests/backends/sql_translator_integration_tests/test_sql_postgres_translator_steps.py
@@ -9,7 +9,6 @@ import psycopg2
 import pytest
 from sqlalchemy import create_engine, text
 from sqlalchemy.engine.base import OptionEngine
-from toucan_connectors.common import nosql_apply_parameters_to_query
 
 from tests.utils import (
     _BEERS_TABLE_COLUMNS,
@@ -21,6 +20,7 @@ from tests.utils import (
 from weaverbird.backends.pypika_translator.dialects import SQLDialect
 from weaverbird.backends.pypika_translator.translate import translate_pipeline
 from weaverbird.pipeline import PipelineWithVariables
+from weaverbird.utils.toucan_connectors import nosql_apply_parameters_to_query_with_errors
 
 con_params = {
     "host": "127.0.0.1",
@@ -75,7 +75,9 @@ def test_sql_translator_pipeline(
 
     steps = spec["step"]["pipeline"]
     steps.insert(0, {"name": "domain", "domain": "beers_tiny"})
-    pipeline = PipelineWithVariables(steps=steps).render(available_variables, nosql_apply_parameters_to_query)
+    pipeline = PipelineWithVariables(steps=steps).render(
+        available_variables, nosql_apply_parameters_to_query_with_errors
+    )
 
     # Convert Pipeline object to Postgres Query
     query = translate_pipeline(