+++ server/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Changed
+- Missing variables during rendering now cause a `UndefinedVariableError`  to be raised instead of a `ValidationError`.
+
 ## [0.51.2] - 2025-02-20
 
 ### Fixed